Removed reliance on ENTRYPOINT in Dockerfiles used by pkgpanda.
This change is in support of running arbitrary pkgpanda scripts to
support testing against generic docker OS images on Windows.

It makes the Linux Dockerfile consistent with the Windows one and
updates the package builder to explicitly invoke the shell (and its
arguments) when launching the docker container backed by those docker
files.