Updated 'pkgpanda/test_http.py' to also copy in the packages directory.
Previously, we were only copying in the 'install' directory, which has
relative symlinks into the 'packages' directory. This means that all of
these symlinks were broken. This does not seem to cause problems on
linux/darwin, but will cause problems on windows. This commit
preemptively fixes this so that we don't run into issues with windows
later on.