Updated 'build_integration_test.py' to run tests on Windows.
This change involved adding a windows variant for all of the package
builds in order to allow it to have a custom docker image in the
'buildinfo.json'. It may be worth looking into how we can build packages
cross platform without requiring a varinat when the only change they
have is which docker container to build in.